backend/requirements-base.txt
pydantic==2.8.2
hyperscan==0.7.9
pyyaml==6.0.2
